Move example to appropriately named folder
The Arduino Library Specification requires that example sketches be located under the examples folder. If you attempt to import a library that does not meet this requirement into the Arduino Web Editor, it will import the example sketch, but not the library.

The Arduino IDE requires the sketch folder name to match the filename of the primary sketch file. This change causes the example sketch to be accessible via the Arduino IDE's File > Examples > LIBRARYNAME menu after the library is installed.

Reference:
https://github.com/arduino/Arduino/wiki/Arduino-IDE-1.5:-Library-specification#library-examples